ratelimitmap: Rename to not be specific to metrics map
The package is currently named ratelimitmetricsmap. A subsequent commit
is going to add a definition for the cilium_ratelimit (i.e., not
metrics) map as well. This renaming is to prepare for that.
