+++ pkg/maps/ratelimitmap/doc.go
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+// Package ratelimitmap represents the BPF ratelimit maps in the BPF programs.
+// It contains stats about BPF ratelimiter.
+package ratelimitmap
