--- pkg/maps/ratelimitmetricsmap/doc.go
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright Authors of Cilium
-
-// Package ratelimitmetricsmap represents the BPF ratelimit metrics map in the BPF programs.
-// It contains stats about BPF ratelimiter.
-package ratelimitmetricsmap
